update for new pydantic version, add caching, split into multiple files
