+++ setup.cfg
@@ -22,7 +22,7 @@ packages =
     yaloader
 python_requires = >=3.9
 install_requires =
-    pydantic
+    pydantic >= 2.0
     pyyaml
 
 [options.extras_require]
