+++ tests/conftest.py
@@ -1,7 +1,8 @@
 import pytest
 
-from yaloader import YAMLConfigLoader, ConfigLoader, YAMLBaseConfig
 import yaloader
+import yaloader.constructor
+from yaloader import YAMLConfigLoader, ConfigLoader, YAMLBaseConfig
 
 
 @pytest.fixture
@@ -14,14 +15,13 @@ def yaml_loader():
 
 @pytest.fixture
 def config_loader(yaml_loader):
-    config_loader = ConfigLoader(yaml_loader=yaml_loader)
+    config_loader = ConfigLoader(yaml_loader=yaml_loader, cacheing=False)
     return config_loader
 
 
 @pytest.fixture
 def AConfig(yaml_loader, config_loader):
-
-    @yaloader.loads(yaml_loader=yaml_loader)
+    @yaloader.constructor.loads(yaml_loader=yaml_loader)
     class Config(YAMLBaseConfig):
         _yaml_tag = '!A'
         attribute: int = 0
