+++ tests/test_overwriting_tags.py
@@ -1,14 +1,14 @@
 import pytest
 
-from yaloader import YAMLBaseConfig
 import yaloader
+import yaloader.constructor
+from yaloader import YAMLBaseConfig
 from yaloader.utils import full_object_name
 
 
 @pytest.fixture
 def BConfig(yaml_loader, AConfig):
-
-    @yaloader.loads(overwrite_tag=True, yaml_loader=yaml_loader)
+    @yaloader.constructor.loads(overwrite_tag=True, yaml_loader=yaml_loader)
     class Config(AConfig):
         _yaml_tag = "!A"
 
@@ -22,9 +22,8 @@ def test_loading_subclass(config_loader, AConfig, BConfig):
 
 
 def test_error_on_standard_tag(yaml_loader):
-
     with pytest.raises(RuntimeError) as error:
-        @yaloader.loads(yaml_loader=yaml_loader)
+        @yaloader.constructor.loads(yaml_loader=yaml_loader)
         class Config(YAMLBaseConfig):
             _yaml_tag = "!!str"
 
@@ -35,7 +34,7 @@ def test_error_on_existing_tag(yaml_loader):
     yaml_loader.add_constructor('!A', lambda _: None)
 
     with pytest.raises(RuntimeError) as error:
-        @yaloader.loads(yaml_loader=yaml_loader)
+        @yaloader.constructor.loads(yaml_loader=yaml_loader)
         class Config(YAMLBaseConfig):
             _yaml_tag = "!A"
 
@@ -43,9 +42,8 @@ def test_error_on_existing_tag(yaml_loader):
 
 
 def test_error_on_registered_tag(yaml_loader, AConfig):
-
     with pytest.raises(RuntimeError) as error:
-        @yaloader.loads(overwrite_tag=False, yaml_loader=yaml_loader)
+        @yaloader.constructor.loads(overwrite_tag=False, yaml_loader=yaml_loader)
         class Config(YAMLBaseConfig):
             _yaml_tag = "!A"
